+++ ui/src/layouts/chats/presentation/chat/mod.rs
@@ -6,12 +6,16 @@ mod pinned_messages;
 mod topbar;
 
 use dioxus::prelude::*;
+use dioxus_html::input_data::keyboard_types::Code;
+use dioxus_html::input_data::keyboard_types::Modifiers;
 
 use kit::{
     components::message_group::MessageGroupSkeletal,
     layout::{modal::Modal, topbar::Topbar},
 };
+use warp::raygun::Location;
 
+use crate::utils::clipboard::clipboard_data::get_files_path_from_clipboard;
 use crate::{
     components::media::calling::CallControl,
     layouts::chats::{
@@ -88,6 +92,38 @@ pub fn Compose(cx: Scope) -> Element {
     cx.render(rsx!(
         div {
             id: "compose",
+            onkeydown: move |e: Event<KeyboardData>| {
+                let keyboard_data = e;
+                let modifiers = if cfg!(target_os = "macos") {
+                    Modifiers::SUPER
+                } else {
+                    Modifiers::CONTROL
+                };
+                if keyboard_data.code() == Code::KeyV
+                    && keyboard_data.modifiers() == modifiers
+                {
+                  let files_local_path = get_files_path_from_clipboard().unwrap_or_default();
+                  if !files_local_path.is_empty() {
+                    let new_files: Vec<Location> = files_local_path
+                    .iter()
+                    .map(|path| Location::Disk { path: path.clone() })
+                    .collect();
+                let mut current_files: Vec<_> = state
+                    .read()
+                    .get_active_chat()
+                    .map(|f| f.files_attached_to_send)
+                    .unwrap_or_default()
+                    .drain(..)
+                    .filter(|x| !new_files.contains(x))
+                    .collect();
+                    current_files.extend(new_files);
+                    let active_chat_id = state.read().get_active_chat().map(|f| f.id).unwrap_or_default();
+                state
+                    .write()
+                    .mutate(Action::SetChatAttachments(active_chat_id, current_files));
+                }
+                }
+            },
             Topbar {
                 with_back_button: state.read().ui.is_minimal_view() && state.read().ui.sidebar_hidden,
                 onback: move |_| {
